# Optimization for special cases in StreamRoute