## Description
Optimize the exact  and the presence match in the case of not inverted rules and single rule per stream or with matching type OR.

## Motivation and Context
The performance of StreamRouteEngine it's O(n), grows with the number of streams. If you have many streams you can process less messages per second that with one stream.

In the case that you have streams with simple rules like exact match or presence we can use a hash table to speed up the process of match this rules.

The patch it's for a special case:
- rule not inverted
- and one of:
   - one rule per stream
   - multiple rules with OR

but i think that is a common case.

The patch may need a rework to fit well in StreamRouteEngine and match/* classes, currently it's a hack to test the concept.
 
## How Has This Been Tested?

To test the improvement in the case of one exact match rule per stream we use the metric: ``org.graylog2.filters.StreamMatcherFilter.executionTime``

Before the patch, the time to match 1, 100 and 1000 streams:
 
- 1 stream

    95th percentile: 21μs
    98th percentile: 25μs
    99th percentile: 36μs
    Standard deviation: 5μs
    Mean: 13μs
    Minimum: 7μs
    Maximum: 75μs

- 100 streams

    95th percentile: 753μs
    98th percentile: 958μs
    99th percentile: 1,088μs
    Standard deviation: 189μs
    Mean: 540μs
    Minimum: 328μs
    Maximum: 3,551μs

- 1000 streams

    95th percentile: 11,558μs
    98th percentile: 16,194μs
    99th percentile: 25,483μs
    Standard deviation: 10,900μs
    Mean: 7,835μs
    Minimum: 4,728μs
    Maximum: 259,153μs

After apply the patch:

-  1000 streams:

    95th percentile: 21μs
    98th percentile: 28μs
    99th percentile: 39μs
    Standard deviation: 5μs
    Mean: 14μs
    Minimum: 9μs
    Maximum: 84μs

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.