@CLAassistant 2017-05-06
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/Graylog2/graylog2-server?pullRequest=3793) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla-assistant.io/Graylog2/graylog2-server?pullRequest=3793) before we can accept your contribution.<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/Graylog2/graylog2-server?pullRequest=3793) it.</sub>