@bernd 2017-09-11
@dennisoelkers Is this ready to be merged?