@bernd 2017-10-03
I am moving this into the 3.0.0 milestone, too late for 2.4.