@dennisoelkers 2018-03-23
Very late response: The code looks good to me and the benefit is without a doubt, but it makes the existing code very complicated for a corner case. I would suggest decoupling the stream router strategy from the router itself and make this a possible option next to the current one, which is either chosen manually by the user or automatically (if the characteristics of the given streams make it viable).