@dennisoelkers 2018-04-05
@manuelluis: Do you have any interesting continuing working on this PR according to my last comment? I would close it otherwise.