@manuelluis 2018-04-05
Yes I'm interested In working in this PR.
I will try to make a proposal as soon as I can.