@dennisoelkers 2019-02-18
Closing this due to inactivity. Please reopen if you want to continue working on it.